# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto de Controle de Filmes - SENAI Jandira
  description: |-
    API destinada para controle de filmes
  termsOfService: http://swagger.io/terms/
  contact:
    email: 09matheushenry@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:3000
paths:
...
  /v1/senai/locadora/diretor:
    post:
      tags:
        - Diretor
      summary: Adiciona um novo Diretor
      description: Adiciona um novo Diretor
      requestBody:
        description: JSON contendo os dados do novo diretor a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Diretor'
            example:
              nome: "Camila Rocha"
              biografia: "Artista"
              data_nascimento: "2009-04-30"
              id_nacionalidade_diretor: 2
              id_sexo_diretor: 2
              foto:
                - id: 2
              atividade:
                - id: 3
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Diretor
      summary: Lista todos os Diretores
      description: Listagem de todos os diretores registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiretorGet'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/diretor/{id}:
    get:
      tags:
        - Diretor
      summary: Busca um Diretor pelo ID
      description: Retorna um Diretor pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do Diretor
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiretorGet'
        '400':
          description: Error Bad Request (ID)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Diretor
      summary: Atualiza um Diretor pelo ID
      description: Atualiza um Diretor pelo ID colocando as informações da mesma forma que o POST
      parameters:
        - name: id
          in: path
          description: ID para atualizar o diretor
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um diretor existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Diretor'
            example:
              nome: "Camila Rocha"
              biografia: "Artista"
              data_nascimento: "2009-04-30"
              id_nacionalidade_diretor: 2
              id_sexo_diretor: 2
              foto:
                - id: 2
              atividade:
                - id: 3
        required: true
      responses:
        '200':
          description: Ok
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Diretor
      summary: Exclui um Diretor pelo ID
      description: Excluir um Diretor existente buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca para a exclusão do diretor
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Success Delete
        '400':
          description: Error Bad Request (ID)
        '500':
          description: Error Internal Server (Controller | Model)

  # ─────────────────────────── ATOR ───────────────────────────
  /v1/senai/locadora/ator:
    post:
      tags:
        - Ator
      summary: Adiciona um novo Ator
      description: Adiciona um novo Ator
      requestBody:
        description: JSON contendo os dados do novo ator a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ator'
            example:
              nome: "Camila Rocha"
              biografia: "Artista"
              data_nascimento: "2009-04-30"
              id_nacionalidade_ator: 2
              id_sexo_ator: 2
              foto:
                - id: 2
              atividade:
                - id: 3
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Ator
      summary: Lista todos os Atores
      description: Listagem de todos os atores registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AtorGet'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/ator/{id}:
    get:
      tags:
        - Ator
      summary: Busca um Ator pelo ID
      description: Retorna um Ator pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do Ator
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtorGet'
        '400':
          description: Error Bad Request (ID)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Ator
      summary: Atualiza um Ator pelo ID
      description: Atualiza um Ator pelo ID colocando as informações da mesma forma que o POST
      parameters:
        - name: id
          in: path
          description: ID para atualizar o ator
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um ator existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ator'
            example:
              nome: "Camila Rocha"
              biografia: "Artista"
              data_nascimento: "2009-04-30"
              id_nacionalidade_ator: 2
              id_sexo_ator: 2
              foto:
                - id: 2
              atividade:
                - id: 3
        required: true
      responses:
        '200':
          description: Ok
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Ator
      summary: Exclui um Ator pelo ID
      description: Excluir um Ator existente buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca para a exclusão do ator
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Success Delete
        '400':
          description: Error Bad Request (ID)
        '500':
          description: Error Internal Server (Controller | Model)

  # ─────────────────────────── GÊNERO ───────────────────────────
...
  /v1/senai/locadora/classificacao:
    post:
      tags:
        - Classificação
      summary: Adiciona uma nova Classificacao
      description: Adiciona uma nova Classificacao
      requestBody:
        description: JSON contendo os dados da nova Classificacao a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Classificacao'
            example:
              classificacao: "+18"
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Classificação
      summary: Lista todas as Classificacoes
      description: Listagem de todas as Classificacoes registradas no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassificacaoGet'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/classificacao/{id}:
    get:
      tags:
        - Classificação
      summary: Busca uma Classificacao pelo ID
      description: Retorna uma Classificacao pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca da Classificacao
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificacaoGet'
        '400':
          description: Error Bad Request (ID)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Classificação
      summary: Atualiza uma Classificacao pelo ID
      description: Atualiza uma Classificacao pelo ID colocando as informações da mesma forma que o POST
      parameters:
        - name: id
          in: path
          description: ID para fazer a busca pela classificacao
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar uma classificacao existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Classificacao'
            example:
              classificacao: "+18"
        required: true
      responses:
        '200':
          description: Ok
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Classificação
      summary: Exclui uma Classificacao pelo ID
      description: Excluir uma Classificacao existente buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca para a exclusão da classificacao
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Success Delete
        '400':
          description: Error Bad Request (ID)
        '500':
          description: Error Internal Server (Controller | Model)

  # ─────────────────────────── SEXO ───────────────────────────
...
  /v1/senai/locadora/atividade:
    post:
      tags:
        - Atividade
      summary: Adiciona uma nova Atividade
      description: Adiciona uma nova Atividade
      requestBody:
        description: JSON contendo os dados da Atividade a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Atividade'
            example:
              atividade: "Dublê"
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Atividade
      summary: Lista todas as Atividades
      description: Listagem de todas as Atividades registradas no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AtividadeGet'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/atividade/{id}:
    get:
      tags:
        - Atividade
      summary: Busca uma Atividade pelo ID
      description: Retorna uma Atividade pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca da Atividade
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtividadeGet'
        '400':
          description: Error Bad Request (ID)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Atividade
      summary: Atualiza uma Atividade pelo ID
      description: Atualiza uma Atividade pelo ID colocando as informações da mesma forma que o POST
      parameters:
        - name: id
          in: path
          description: ID para fazer a busca pela atividade
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar uma atividade existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Atividade'
            example:
              atividade: "Dublê"
        required: true
      responses:
        '200':
          description: Ok
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Atividade
      summary: Exclui uma Atividade pelo ID
      description: Excluir uma Atividade existente buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca para a exclusão da atividade
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Success Delete
        '400':
          description: Error Bad Request (ID)
        '500':
          description: Error Internal Server (Controller | Model)

  # ─────────────────────────── FOTO ───────────────────────────
...
components:
  schemas:
...
    IdRef:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          example: 1

    # ── FILME ──
...
    Diretor:
      required:
        - nome
        - data_nascimento
        - id_sexo_diretor
        - id_nacionalidade_diretor
      type: object
      properties:
        nome:
          type: string
          maxLength: 100
          example: "Camila Rocha"
          description: Máximo de 100 caracteres
        biografia:
          type: string
          example: "Artista"
          description: Sem limite de caracteres
        data_nascimento:
          type: string
          format: date
          example: "2009-04-30"
        id_sexo_diretor:
          type: integer
          format: int64
          example: 2
          description: ID do sexo do diretor
        id_nacionalidade_diretor:
          type: integer
          format: int64
          example: 2
          description: ID da nacionalidade do diretor
        foto:
          type: array
          description: Lista de IDs de fotos associadas ao diretor
          items:
            $ref: '#/components/schemas/IdRef'
          example:
            - id: 2
        atividade:
          type: array
          description: Lista de IDs de atividades exercidas pelo diretor
          items:
            $ref: '#/components/schemas/IdRef'
          example:
            - id: 3
    DiretorGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          maxLength: 100
          example: "Camila Rocha"
        biografia:
          type: string
          example: "Artista"
        data_nascimento:
          type: string
          format: date
          example: "2009-04-30"
        id_sexo_diretor:
          type: integer
          format: int64
          example: 2
        id_nacionalidade_diretor:
          type: integer
          format: int64
          example: 2
        foto:
          type: array
          items:
            $ref: '#/components/schemas/IdRef'
        atividade:
          type: array
          items:
            $ref: '#/components/schemas/IdRef'

    # ── ATOR ──
    Ator:
      required:
        - nome
        - data_nascimento
        - id_sexo_ator
        - id_nacionalidade_ator
      type: object
      properties:
        nome:
          type: string
          maxLength: 100
          example: "Camila Rocha"
          description: Máximo de 100 caracteres
        biografia:
          type: string
          example: "Artista"
          description: Sem limite de caracteres
        data_nascimento:
          type: string
          format: date
          example: "2009-04-30"
        id_sexo_ator:
          type: integer
          format: int64
          example: 2
          description: ID do sexo do ator
        id_nacionalidade_ator:
          type: integer
          format: int64
          example: 2
          description: ID da nacionalidade do ator
        foto:
          type: array
          description: Lista de IDs de fotos associadas ao ator
          items:
            $ref: '#/components/schemas/IdRef'
          example:
            - id: 2
        atividade:
          type: array
          description: Lista de IDs de atividades exercidas pelo ator
          items:
            $ref: '#/components/schemas/IdRef'
          example:
            - id: 3
    AtorGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          maxLength: 100
          example: "Camila Rocha"
        biografia:
          type: string
          example: "Artista"
        data_nascimento:
          type: string
          format: date
          example: "2009-04-30"
        id_sexo_ator:
          type: integer
          format: int64
          example: 2
        id_nacionalidade_ator:
          type: integer
          format: int64
          example: 2
        foto:
          type: array
          items:
            $ref: '#/components/schemas/IdRef'
        atividade:
          type: array
          items:
            $ref: '#/components/schemas/IdRef'

    # ── GÊNERO ──
...
    Classificacao:
      required:
        - classificacao
      type: object
      properties:
        classificacao:
          type: string
          maxLength: 6
          example: "+18"
          description: Máximo de 6 caracteres
    ClassificacaoGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        classificacao:
          type: string
          maxLength: 6
          example: "+18"

    # ── SEXO ──
...
    Atividade:
      required:
        - atividade
      type: object
      properties:
        atividade:
          type: string
          maxLength: 40
          example: "Dublê"
          description: Máximo de 40 caracteres
    AtividadeGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        atividade:
          type: string
          maxLength: 40
          example: "Dublê"

    # ── FOTO ──
...